% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{The child's game Fizz Buzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{An integer number}
}
\value{
The vector of \code{n} as \code{fizzList}
}
\description{
Function that outputs a vector of first n terms in the Child's game
Fizz Buzz.
}
\examples{
FizzBuzz(12)
FizzBuzz(4)
}
