% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily Flagstaff Weather from 2015-2019.}
\format{
a data frame of 1719 observations
\describe{
   \item{DATE}{The date of observation as a POSIX date format.}
   \item{PRCP}{Precipitation}
   \item{SNOW}{Snowfall}
   \item{TMAX}{Maximum temperature}
}
}
\source{
\file{Pulliam_Airport_Weather_Station.csv}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily Flagstaff Weather from 2015-2019.
}
\keyword{datasets}
